+++ pkg/sdkv2/resources/job.go
@@ -65,7 +65,7 @@ var jobSchema = map[string]*schema.Schema{
 		Type:        schema.TypeBool,
 		Optional:    true,
 		Default:     true,
-		Description: "Flag for whether the job is marked active or deleted. To create/keep a job in a 'deactivated' state, check  the `triggers` config.",
+		Description: "Should always be set to true as setting it to false is the same as creating a job in a deleted state. To create/keep a job in a 'deactivated' state, check  the `triggers` config.",
 	},
 	"triggers": &schema.Schema{
 		Type:     schema.TypeMap,